Project: Restructured Architecture to accomodate frontend

+++ backend/contracts/Compound/DefiProject.sol
@@ -15,17 +15,17 @@ contract DefiProject {
         priceOracle = PriceOracleInterface(_priceOracle);
     }
 
-    function supply(address cTokenAddress, uint256 underlyingAmount) external {
+    function supply(address cTokenAddress, uint underlyingAmount) external {
         CTokenInterface cToken = CTokenInterface(cTokenAddress);
         address underlyingAddress = cToken.underlying();
         IERC20(underlyingAddress).approve(cTokenAddress, underlyingAmount);
-        uint256 result = cToken.mint(underlyingAmount);
+        uint result = cToken.mint(underlyingAmount);
         require(result == 0, "cToken mint failed, see Compound ErrorReporter.sol for more details");
     }
 
-    function redeem(address cTokenAddress, uint256 cTokenAmount) external {
+    function redeem(address cTokenAddress, uint cTokenAmount) external {
         CTokenInterface cToken = CTokenInterface(cTokenAddress);
-        uint256 result = cToken.redeem(cTokenAmount);
+        uint result = cToken.redeem(cTokenAmount);
         require(
             result == 0,
             "cToken redeem failed, see Compound ErrorReporter.sol for more details"
@@ -35,35 +35,35 @@ contract DefiProject {
     function enterMarket(address cTokenAddress) external {
         address[] memory markets = new address[](1);
         markets[0] = cTokenAddress;
-        uint256[] memory results = comptroller.enterMarkets(markets);
+        uint[] memory results = comptroller.enterMarkets(markets);
         require(
             results[0] == 0,
             "comptroller enterMarket failed, see Compound ErrorReporter.sol for more details"
         );
     }
 
-    function borrow(address cTokenAddress, uint256 borrowAmount) external {
+    function borrow(address cTokenAddress, uint borrowAmount) external {
         CTokenInterface cToken = CTokenInterface(cTokenAddress);
         address underlyingAddress = cToken.underlying();
-        uint256 result = cToken.borrow(borrowAmount);
+        uint result = cToken.borrow(borrowAmount);
         require(
             result == 0,
             "cToken borrow failed, see Compound ErrorReporter.sol for more details"
         );
     }
 
-    function repayBorrow(address cTokenAddress, uint256 underlyingAmount) external {
+    function repayBorrow(address cTokenAddress, uint underlyingAmount) external {
         CTokenInterface cToken = CTokenInterface(cTokenAddress);
         address underlyingAddress = cToken.underlying();
         IERC20(underlyingAddress).approve(cTokenAddress, underlyingAmount);
-        uint256 result = cToken.repayBorrow(underlyingAmount);
+        uint result = cToken.repayBorrow(underlyingAmount);
         require(
             result == 0,
             "cToken repayBorrow failed, see Compound ErrorReporter.sol for more details"
         );
     }
 
-    function getMaxBorrow(address cTokenAddress) external view returns (uint256) {
+    function getMaxBorrow(address cTokenAddress) external view returns (uint) {
         (uint result, uint liquidity, uint shortfall) = comptroller.getAccountLiquidity(
             address(this)
         );
@@ -73,7 +73,7 @@ contract DefiProject {
         );
         require(shortfall == 0, "Account underwater...");
         require(liquidity != 0, "Account doesn't have any collateral");
-        uint256 underlyingPrice = priceOracle.getUnderlyingPrice(cTokenAddress);
+        uint underlyingPrice = priceOracle.getUnderlyingPrice(cTokenAddress);
         return liquidity / underlyingPrice;
     }
 }
