+++ backend/contracts/Compound/PriceOracleInterface.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.7;
 
 interface PriceOracleInterface {
-    function getUnderlyingPrice(address cTokenAddress) external view returns (uint256);
+    function getUnderlyingPrice(address cTokenAddress) external view returns (uint);
 }
